FCalcDescr now shouldn't fail on small inputs
The routines in c2xDistr's FCalcDescr and its subclasses sometimes assume the
number of frames of analysis data to be greater than 1, which causes failure
when this is not the case. A new function FCalcDescrCheckFields has been added
to c2xDistr and its subclasses (when they need to override the default
functionality). The fields can be checked for a variety of things, currently all
the implementations (that of c2xDistr and its subclasses) check to see that the
number of frames of the analysis data be greater than 1. If not, an error is
thrown that can be caught.

The catching of these errors is used in Gget_desc_onefile_do_by_chunks.m as one
of the idicators of when to stop analysing the sound. It has been tested on a
file that was sized in such a way that the size of the last chunk was around 450
samples: smaller that the window size of any analysis. These analyses would
return no frames or 1 frame. If this is the case, the descriptors are not
calculated and the calculation is complete for that file.